Clean up the occurrence string presented to the user
If `iedit-only-complete-symbol-*' is being used then there are some
extra characters added to the occurrence string which make the string
a regexp to only match symbols. The user doesn't need to know what the
regexp looks like, they only care about the occurrence string, so
clean up the added characters if necessary when presenting the
occurrence string to the user.