Option to clean up the occurrence string presented to the user
If `iedit-only-complete-symbol-*' is being used then there are some
extra characters added to the occurrence string which make the string
a regexp to only match symbols. The user might not want to see the
extra characters needed to form the regexp. Provide an option to strip
off the added characters when presenting the occurrence string to the
user.