Add a new feature - preserve case
The behavior of this feature is the same as `query-replace'.
If `iedit-case-sensitive' is nil and `case-replace' is non-nil, iedit will
preserve case in occurrences.

Quoted from doc-string of `replace-match':

"maybe capitalize the whole text, or maybe just word initials, based on the
replaced text.  If the replaced text has only capital letters and has at least
one multiletter word, convert NEWTEXT to all caps.  Otherwise if all words are
capitalized in the replaced text, capitalize each word in NEWTEXT."