Separate iedit setup into `iedit-begin'
`iedit-start' consists of two parts: scanning for a regex, and setting-up
various hooks so that `iedit' works correctly (`iedit-done', ...).

The latter part is useful in isolation if there is a custom way to acquire
`iedit-occurrences-overlays', as is the case with `lsp-mode', which recently got
`iedit' integration. This way, the `add-hook' blocks no longer need to be
duplicated in such packages.