Simplify iedit-replace-occurrences and iedit-apply-on-occurrences
* `iedit-apply-on-occurrences' no longer accept extra arguments; it only
  expects a function that will accept 2 arguments: the begin and end position
  of the occurrence it must process.  The current implementation does not have
  a need for extra arguments, except the lambda passed by
  `iedit-replace-occurrences'
  but that can change:
* `iedit-replace-occurrences' is also simplified.  The lambda passed is also a
  closure that remembers the original string and its replacement, there's no
  need to pass these variables to `iedit-apply-on-occurrences' as it was done.

`iedit-apply-on-occurrences' no longer uses apply since it does not accept a
optional list of extra arguments; it uses funcall instead now that it does not
have to deal with a potential extra argument list.

Overall, this change is a cleanup optimization.  The old code was using a
closure and yet passed the same value through arguments.
That's no longer the case.