Integrate Victor Ren changes in iedit-apply-on-occurrences-test
With additional indentation changes and string format to keep the visual
format of the string that will be inserted in the test buffer.
To keep this visual clue while respecting the code indentation, the string
opening double quote must be followed by a backslash and the text can then
begin on the next line the way it ill show up in the test buffer.